+++ Library/trienode.cpp
@@ -10,7 +10,7 @@ trienode::trienode()
 	firstoffset(0),
 	letter('\0'),
 	wordend(false)
-{for (int i = 0; i < LINKS_LENGTH; ++i) links[i] = NULL_LINK;}
+{for (long i = 0; i < LINKS_LENGTH; ++i) links[i] = NULL_LINK;}
 
 trienode::trienode(long _firstoffset,
 	long _nrofoccurences, unsigned char _letter,
@@ -20,11 +20,11 @@ trienode::trienode(long _firstoffset,
 	nrofoccurences(_nrofoccurences),
 	letter(_letter),
 	wordend(_wordend)
-{for (int i = 0; i < LINKS_LENGTH; ++i) links[i] = NULL_LINK;}
+{for (long i = 0; i < LINKS_LENGTH; ++i) links[i] = NULL_LINK;}
 
 long trienode::operator[](char c)
 {
-	return links[(int) c];
+	return links[(long) c];
 }
 
 void trienode::set_link(char idx, long val)
@@ -42,7 +42,7 @@ printf(" # offset: %-6i#\n",firstoffset);
 printf(" # count:  %-6i#\n",(nrofoccurences <= 0? 0:nrofoccurences));
 printf(" # end?:   %-6i#\n",wordend);
 printf(" #---------------#\n");
-for(int j=0 ; j<256;j++)
+for(long j=0 ; j<256;j++)
 if(links[j]!=-1)
 printf(" #%6c->%-6i #\n",j,links[j]);
 printf(" #################\n");
